Initial porting of IaC from CDK to Pulumi
Comsidering Pulumi may become the official IaC of Jupiter in the future, I have decided to implement it for
this workshop in order to start testing/using it.

My goal is not to replace CDK, rather I aim to create an alternative infrastructure as code option.
This provides flexibility in choosing between CDK and Pulumi for future developments.