Porting Pulumi as a second deploying system for the commit workshop
Porting is complete: the purpose was to create the same implementation as the one provided by CDK, but on Pulumi.
Both pulumi and CDK should work fine and it is possible to choose at beginning which one to use (CDK should be tested though).
ReadMe is updated with a tutorial on how to setup the workshop.